Formal traits for to_index (#14)
* Implement Proper Key-Indices Trait system

This is the first attempt at a formal trait system for keys and indices
based indexing. There are probably some breaking changes but I've mostly
been worried about getting passing tests with a reasonable design.

* Document ToIndexStyle

* Minor fixes to docs (typos, organization)

* Remove dead traits code

* Fix comments on mapped arrays

* Remove unecessary boundschecks

AbstractAxis doesn't need boundscheck at getindex b/c it already has
this at the to_index level

* Reorganize code and docs

* Add tests for Julia v1.3

* Anonymous isapprox support

This is not final at all. This implementation scares and I'd like
something that more safely refers to the exact anonymous function being
used. However, it seems to work...for now

* Fix 1.2 failures

This forced a small rewrite of this branch due to a corner case with
boolean indexing.

* Clean up old unecessary code

* Missed this test

* Re-cover proper error tests

* Delete empty file

* Fix inference issues

* More inference tests + v0.2.4