Incorporating new ArrayInterface indexing functionality
This is a huge rewrite of the back end, that makes it generic to how
other arrays are indexed using ArrrayInterface. This allows us to
support things like static sizes without depending directly on
StaticArrays and will hopefully result in compatability with
LoopVectorization in the not too distant future.

This also enforces axis arguments to match the number of dimensions for
the parent array. Now we assume that unit ranges that aren't one based
are intended to be offset axes.

Methods that involve combining multiple axes (cat/broadcasting/etc) try
to account for the possibility of an `AbstractAxis` type wrapping
another axis type.